Treat calories as int

+++ src/YaleMenus/Models.swift
@@ -184,7 +184,7 @@ extension Item: Decodable {
 
 struct Nutrition {
     let servingSize: String?
-    let calories: Double
+    let calories: Int
 
     let totalFat: String
     let saturatedFat: String
@@ -265,7 +265,7 @@ extension Nutrition: Decodable {
         let container = try decoder.container(keyedBy: NutritionCodingKeys.self)
 
         servingSize = try container.decode(String?.self, forKey: .servingSize)
-        calories = try container.decode(Double.self, forKey: .calories)
+        calories = try container.decode(Int.self, forKey: .calories)
 
         totalFat = try container.decode(String.self, forKey: .totalFat)
         saturatedFat = try container.decode(String.self, forKey: .saturatedFat)
